* Modified the configure.ac and the src/Makefile.am so that when built for windows using mingw it would link to the ws2_32 (a.k.a. Winsock 2 32-bit library).
* Modified the scripts/build_sxs_error_files.py script to include
warnings in both the sxs_error.h and sxs_error.c files reminding
developers not to edit them by hand since they are autogenerated and
their modifications will be lost.

* Modified the unix_errs.in by placing the ECANCELED error in it.
Modified the win_errs.in by placing the WSAECANCELLED error in it. This
was necessary because in windows they use the double L spelling of the
word canceled while on Unix systems they use the single L version. It
seems that it was a typo on Windows part because in the Windows
documentation they describe the error using the single L spelling of the
word yet the double LL spelling for the macro. Anyways, I also removed
the ECANCELED error from the both_errs.in because it wouldn't compile
when building for windows due to it not being defined because of the
spelling miss-match.

* Modified the README file by cleaning it up and adding a section
covering Producing a MSVC Compatible Build, Using MSVC to Link to MSVC
Compatible Build, and I extended the Building section to clear up
things. I also corrected the Directory Contents sections.