Refixed Dave's breaking of the universal build.
Dave for some reason or another decided that a library which was
designed to be cross-platform and buit as universal Mac OS X binary for
Mac OS X 10.4 and 10.5 should no longer be built as a universal binary
and instead only as a 10.5 binary. Hence, I reworked the configure.ac so
that building it as a universal binary for 10.4 and 10.5 is an option to
the configure script and that by default it builds a 10.5 binary only.